fix: remove require fields modal
