+++ languages/woo-pagarme-payments-pt_BR.po
@@ -5,7 +5,8 @@ msgstr ""
 "Project-Id-Version: WooCommerce Pagar.me Payments 1.0\n"
 "Report-Msgid-Bugs-To: https://wordpress.org/support/plugin/woo-pagarme-payments\n"
 "POT-Creation-Date: 2018-06-22 13:58-0300\n"
-"PO-Revision-Date: 2024-03-20 14:27-0300\n"
+"PO-Revision-Date: 2024-03-21 17:38-0300\n"
+"PO-Revision-Date: 2024-03-21 16:52-0300\n"
 "Last-Translator: Pagar.me\n"
 "Language-Team: \n"
 "Language: pt_BR\n"
@@ -1347,3 +1348,6 @@ msgstr "N - Transação rejeitada pelo Emissor"
 
 msgid "R - Transaction rejected by Issuer (post-challenge)"
 msgstr "R - Transação rejeitada pelo Emissor (pós-desafio)"
+
+msgid "Invalid Data"
+msgstr "Dados Inválidos"
