+++ src/Block/Checkout/Form/Card.php
@@ -11,10 +11,11 @@ declare( strict_types=1 );
 
 namespace Woocommerce\Pagarme\Block\Checkout\Form;
 
-use Woocommerce\Pagarme\Block\Checkout\Gateway;
+use Woocommerce\Pagarme\Core;
 use Woocommerce\Pagarme\Helper\Utils;
-use Woocommerce\Pagarme\Model\Payment\Voucher;
 use Woocommerce\Pagarme\Model\Subscription;
+use Woocommerce\Pagarme\Model\Payment\Voucher;
+use Woocommerce\Pagarme\Block\Checkout\Gateway;
 
 defined( 'ABSPATH' ) || exit;
 
@@ -24,6 +25,8 @@ defined( 'ABSPATH' ) || exit;
  */
 class Card extends Gateway
 {
+    const INVALID_CARD_ERROR_MESSAGE = 'This card number is invalid.';
+    
     /**
      * @var string
      */
@@ -37,11 +40,22 @@ class Card extends Gateway
     /**
      * @var string[]
      */
-    protected $deps = ['pagarme-checkout-card'];
+    protected $deps = [];
 
     /** @var int  */
     protected $sequence = 1;
 
+    public function enqueue_scripts($scripts = null, $deps = [])
+    {
+        parent::enqueue_scripts($scripts, $deps);
+
+        wp_localize_script(
+            WCMP_JS_HANDLER_BASE_NAME . 'card',
+            'PagarmeGlobalVars',
+            self::getLocalizeScriptArgs()
+        );
+    }
+    
     /**
      * @param int $sequence
      * @return $this
@@ -133,4 +147,64 @@ class Card extends Gateway
         return Utils::get_component('pagarme-checkout');
     }
 
+    public static function getLocalizeScriptArgs($args = array())
+    {
+        $defaults = array(
+            'ajaxUrl'        => Utils::get_admin_url('admin-ajax.php'),
+            'WPLANG'         => get_locale(),
+            'spinnerUrl'     => Core::plugins_url('assets/images/icons/spinner.png'),
+            'prefix'         => Core::PREFIX,
+            'checkoutErrors' => array(
+                'pt_BR' => self::getCardErrorsMessagesTranslated(),
+            ),
+        );
+
+        return array_merge($defaults, $args);
+    }
+
+    public static function getCardErrorsMessagesTranslated()
+    {
+        return array(
+            'exp_month: A value is required.'                             =>
+                __('Expiration Date: The month is required.', 'woo-pagarme-payments'),
+            'exp_month: The field exp_month must be between 1 and 12.'    =>
+                __('Expiration Date: The month must be between 1 and 12.', 'woo-pagarme-payments'),
+            "exp_year: The value 'undefined' is not valid for exp_year."  =>
+                __('Expiration Date: Invalid year.', 'woo-pagarme-payments'),
+            'request: The card expiration date is invalid.'               =>
+                __('Expiration Date: Invalid expiration date.', 'woo-pagarme-payments'),
+            'request: Card expired.'                                      =>
+                __('Expiration Date: Expired card.', 'woo-pagarme-payments'),
+            'holder_name: The holder_name field is required.'             =>
+                __('The card holder name is required.', 'woo-pagarme-payments'),
+            'number: The number field is required.'                       =>
+                __('The card number is required.', 'woo-pagarme-payments'),
+            'number: The number field is not a valid credit card number.' =>
+                __(self::INVALID_CARD_ERROR_MESSAGE, 'woo-pagarme-payments'),
+            'card: The number field is not a valid card number'           =>
+                __(self::INVALID_CARD_ERROR_MESSAGE, 'woo-pagarme-payments'),
+            'card.number: The field number must be a string with a minimum length of 13 and a maximum length of 19.'
+            => __('The card number must be between 13 and 19 characters.', 'woo-pagarme-payments'),
+            'card: Card expired.'                                         =>
+                __('The expiration date is expired.', 'woo-pagarme-payments'),
+            'card.cvv: The field cvv must be a string with a minimum length of 3 and a maximum length of 4.'
+            => __('The card code must be between 3 and 4 characters.', 'woo-pagarme-payments'),
+            'card: Invalid data to change card brand'                     =>
+                __(self::INVALID_CARD_ERROR_MESSAGE, 'woo-pagarme-payments'),
+            'card: Tokenize timeout'                                      =>
+                __('Timeout na tokenização.', 'woo-pagarme-payments'),
+            'fail_get_token'        =>
+                __('Failed to generate Token for 3ds, try again.', 'woo-pagarme-payments'),
+            'fail_get_email'        =>
+                __('There was a problem finding the email.', 'woo-pagarme-payments'),
+            'fail_get_billing_address'        =>
+                __('There was a problem finding the addresses.', 'woo-pagarme-payments'),
+            'fail_assemble_card_expiry_date'        =>
+                __('There was a problem when assembling the card\'s expiration data.', 'woo-pagarme-payments'),
+            'fail_assemble_purchase'        =>
+                __('There was a problem when assembling the purchase data.', 'woo-pagarme-payments'),
+            'invalidBrand'        =>
+                __('Invalid Data', 'woo-pagarme-payments'),
+        );
+    }
 }
